+++ cmd/opentelemetry-prometheus-sidecar/main_test.go
@@ -20,10 +20,12 @@ import (
 	"net/http"
 	"os"
 	"os/exec"
+	"sync"
 	"testing"
 	"time"
 
 	"github.com/go-kit/kit/log"
+	traces "github.com/lightstep/opentelemetry-prometheus-sidecar/internal/opentelemetry-proto-gen/trace/v1"
 	"github.com/stretchr/testify/require"
 )
 
@@ -63,19 +65,6 @@ func TestStartupInterrupt(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping test in short mode.")
 	}
-
-	ts := newTestServer(t)
-
-	runPrometheusService(ts)
-	go func() {
-		// By sleeping 5 seconds here, we require the
-		// sidecar's selftest to try and fail for 5 seconds
-		// before succeeding, after which the interrupt is
-		// delivered here.
-		time.Sleep(5 * time.Second)
-		runMetricsService(ts)
-	}()
-	defer ts.Stop()
 
 	cmd := exec.Command(
 		os.Args[0],
@@ -126,30 +115,27 @@ Loop:
 		time.Sleep(time.Second)
 	}
 
-	t.Logf("stdout: %v\n", bout.String())
-	t.Logf("stderr: %v\n", berr.String())
 	if !startedOk {
 		t.Errorf("opentelemetry-prometheus-sidecar didn't start in the specified timeout")
 		return
 	}
 	if err := cmd.Process.Kill(); err == nil {
-		t.Errorf("opentelemetry-prometheus-sidecar didn't shutdown gracefully after sending the Interrupt signal")
-	} else if stoppedErr != nil && stoppedErr.Error() != "signal: interrupt" {
-		// TODO - find a better way to detect when the process didn't exit as expected!
-		// (See *os.ProcessState)
-		t.Errorf("opentelemetry-prometheus-sidecar exited with an unexpected error:%v", stoppedErr)
+		t.Errorf("opentelemetry-prometheus-sidecar didn't shutdown after sending the Interrupt signal")
 	}
+	const expected = "source is not ready: context canceled"
+	require.Error(t, stoppedErr)
+	require.Contains(t, stoppedErr.Error(), "exit status 1")
 
 	// Because the fake endpoint was started after the start of
 	// the test, we should see some gRPC warnings the connection up
 	// until --startup.timeout takes effect.
-	require.Contains(t, berr.String(), "connect: connection refused")
+	require.Contains(t, berr.String(), expected)
 
 	// The process should have been interrupted.
 	require.Contains(t, berr.String(), "received SIGTERM, exiting")
 
-	// The selftest should have finished, since we waited for ready.
-	require.Contains(t, berr.String(), "selftest was successful")
+	// The selftest was interrupted.
+	require.Contains(t, berr.String(), "selftest failed, not starting")
 }
 
 func TestMainExitOnFailure(t *testing.T) {
@@ -243,4 +229,115 @@ func TestStartupUnhealthyEndpoint(t *testing.T) {
 
 	require.Contains(t, berr.String(), "selftest failed, not starting")
 	require.Contains(t, berr.String(), "selftest recoverable error, still trying")
+}
+
+func TestSuperStackDump(t *testing.T) {
+	// Tests that the selftest detects an unhealthy endpoint during the selftest.
+	if testing.Short() {
+		t.Skip("skipping test in short mode.")
+	}
+
+	cmd := exec.Command(
+		os.Args[0],
+		append(e2eTestMainSupervisorFlags,
+			"--prometheus.wal=testdata/wal",
+			"--healthcheck.period=1s",
+			"--startup.delay=0s",
+			"--log.level=debug",
+		)...)
+
+	ms := newTestServer(t)
+	go runMetricsService(ms)
+	runPrometheusService(ms) // Note: there's no metadata api here, we'll see failures.
+
+	ts := newTraceServer(t)
+	go runDiagnosticsService(ms, ts)
+
+	cmd.Env = append(os.Environ(), "RUN_MAIN=1")
+	var bout, berr bytes.Buffer
+	cmd.Stdout = &bout
+	cmd.Stderr = &berr
+	err := cmd.Start()
+	if err != nil {
+		t.Errorf("execution error: %v", err)
+		return
+	}
+
+	var lock sync.Mutex
+	var diagSpans []*traces.ResourceSpans
+
+	ctx, cancel := context.WithCancel(context.Background())
+	var wg sync.WaitGroup
+	wg.Add(1)
+
+	go func() {
+		defer wg.Done()
+		lock.Lock()
+		defer lock.Unlock()
+		for {
+			select {
+			case rs := <-ts.spans:
+				// Note: below searching for a stack dump and
+				// a few key strings, as a simple test.  TODO:
+				// improve by testing support, factoring the
+				// special-purpose logic here into another
+				// package.
+				diagSpans = append(diagSpans, rs)
+			case <-ctx.Done():
+				return
+			}
+		}
+	}()
+
+	go func() {
+		for _ = range ms.metrics {
+			// Dumping the metrics diagnostics here.
+			// TODO: add testing support to validate
+			// metrics from tests and then build more
+			// tests based on metrics.
+		}
+	}()
+
+	// The process is expected to kill itself.
+	err = cmd.Wait()
+	require.Error(t, err)
+
+	cancel()
+	wg.Wait()
+	ms.Stop()
+	ts.Stop()
+
+	lock.Lock()
+	defer lock.Unlock()
+
+	foundCrash := false
+	for _, rs := range diagSpans {
+		for _, span := range rs.InstrumentationLibrarySpans[0].Spans {
+			require.Contains(t, []string{"health-client", "shutdown-report"}, span.Name)
+
+			if span.Name == "shutdown-report" {
+				continue
+			}
+			sa := map[string]string{}
+			for _, a := range span.Attributes {
+				sa[a.Key] = a.Value.String()
+			}
+			statusCode := sa["http.status_code"]
+			require.Contains(t, []string{
+				"int_value:200 ",
+				"int_value:503 ",
+			}, statusCode)
+
+			if statusCode == "int_value:200 " {
+				continue
+			}
+
+			foundCrash = true
+			require.Contains(t, sa["sidecar.status"], "unhealthy")
+			require.Contains(t, sa["sidecar.stackdump"], "goroutine")
+			require.Contains(t, sa["sidecar.stackdump"], "net/http.(*ServeMux).ServeHTTP")
+		}
+	}
+
+	require.True(t, foundCrash, "expected to find a crash report")
 }