+++ supervisor/supervisor.go
@@ -49,26 +49,27 @@ var (
 )
 
 const (
-	healthURI = "/-/health?supervisor=true"
-	readyURI  = "/-/ready"
-
 	supervisorBufferSize = 1 << 14
 )
 
 type (
 	Config struct {
-		Logger log.Logger
-		Admin  config.AdminConfig
-		Period time.Duration
+		Logger    log.Logger
+		Admin     config.AdminConfig
+		Telemetry *telemetry.Telemetry
 	}
 
 	Supervisor struct {
 		logger   log.Logger
 		endpoint string
-		period   time.Duration
 		client   *http.Client
-		readied  bool
+		telem    *telemetry.Telemetry
+		period   time.Duration
 
+		// isRunning means the sidecar enter its run state.
+		isRunning bool
+
+		// veryUnhealthy means the sidecar returned a stackdump.
 		veryUnhealthy bool
 
 		lock      sync.Mutex
@@ -86,9 +87,10 @@ func New(cfg Config) *Supervisor {
 	}
 	return &Supervisor{
 		logger:    cfg.Logger,
-		period:    cfg.Period,
 		endpoint:  endpoint,
+		period:    cfg.Admin.HealthCheckPeriod.Duration,
 		client:    client,
+		telem:     cfg.Telemetry,
 		logBuffer: make([]string, 0, config.DefaultSupervisorLogsHistory),
 	}
 }
@@ -154,10 +156,11 @@ func (s *Supervisor) supervise(ctx context.Context, cmd *exec.Cmd, wg *sync.Wait
 	defer wg.Done()
 
 	for {
-		sleep := s.period
-		ready, _ := s.checkTarget()
-		if !ready {
-			sleep = config.DefaultHealthCheckTimeout
+		var sleep time.Duration
+		if s.isRunning {
+			sleep = s.period
+		} else {
+			sleep = config.DefaultReadinessPeriod
 		}
 
 		// Sleep or be canceled.
@@ -177,9 +180,17 @@ func (s *Supervisor) supervise(ctx context.Context, cmd *exec.Cmd, wg *sync.Wait
 		s.healthcheck(ctx)
 
 		if s.veryUnhealthy {
+			// A stackdump was already reported. We're
+			// going to flush these spans, stop reporting
+			// telemetry, then kill the process.
+			s.finalSpan(errors.New("process is unhealthy"))
+
+			s.telem.Shutdown(ctx)
+
 			// The process is repeatedly failing its internal health check,
 			// and we have seen a stackdump already.
 			cmd.Process.Kill()
+			return
 		}
 	}
 }
@@ -195,8 +206,6 @@ func (s *Supervisor) healthcheck(ctx context.Context) {
 func (s *Supervisor) healthcheckErr(ctx context.Context) (err error) {
 	ctx, cancel := context.WithTimeout(ctx, config.DefaultHealthCheckTimeout)
 	defer cancel()
-
-	ready, target := s.checkTarget()
 
 	ctx, span := tracer.Start(ctx, "health-client")
 
@@ -215,7 +224,7 @@ func (s *Supervisor) healthcheckErr(ctx context.Context) (err error) {
 
 	// Make the request and try to parse the result.
 	err = func() error {
-		req, err := http.NewRequestWithContext(ctx, "GET", target, nil)
+		req, err := http.NewRequestWithContext(ctx, "GET", s.endpoint+config.HealthCheckURI, nil)
 		if err != nil {
 			return errors.Wrap(err, "build request")
 		}
@@ -226,6 +235,8 @@ func (s *Supervisor) healthcheckErr(ctx context.Context) (err error) {
 		}
 		defer resp.Body.Close()
 
+		// Note: The Go SDK 0.16 appears to lose track of the codes.Error status
+		// being set here. It does not appear in the output span.  TODO: investigate.
 		span.SetAttributes(semconv.HTTPAttributesFromHTTPStatusCode(resp.StatusCode)...)
 		span.SetStatus(semconv.SpanStatusFromHTTPStatusCode(resp.StatusCode))
 
@@ -253,18 +264,11 @@ func (s *Supervisor) healthcheckErr(ctx context.Context) (err error) {
 		return nil
 	}()
 
-	good := err == nil
-
 	var hstat string
-	switch {
-	case ready && good:
+	if err == nil {
 		hstat = s.noteHealthy(hr)
-	case ready && !good:
+	} else {
 		hstat = s.noteUnhealthy(hr)
-	case !ready && good:
-		hstat = s.noteReady()
-	case !ready && !good:
-		hstat = s.noteUnready()
 	}
 
 	span.SetAttributes(label.String("sidecar.health", hstat))
@@ -281,7 +285,7 @@ func (s *Supervisor) finalSpan(err error) {
 
 	if err != nil {
 		span.RecordError(err)
-		span.SetStatus(codes.Error, "unexpected shutdown")
+		span.SetStatus(codes.Error, err.Error())
 		return
 	}
 	span.SetStatus(codes.Ok, "graceful shutdown")
@@ -375,38 +379,19 @@ func (s *Supervisor) copyLogs() []string {
 	return cp
 }
 
-func (s *Supervisor) checkTarget() (bool, string) {
-	s.lock.Lock()
-	defer s.lock.Unlock()
-
-	if s.readied {
-		return true, s.endpoint + healthURI
-	}
-	return false, s.endpoint + readyURI
-}
-
-func (s *Supervisor) noteReady() string {
-	level.Info(s.logger).Log("msg", "sidecar is ready")
-
-	s.lock.Lock()
-	defer s.lock.Unlock()
-
-	s.readied = true
-	return "first time ready"
-}
-
-func (s *Supervisor) noteUnready() string {
-	level.Info(s.logger).Log("msg", "sidecar is not ready")
-	return "not ready yet"
-}
-
 func (s *Supervisor) noteHealthy(hr health.Response) string {
-	summary := []interface{}{
-		"msg", "sidecar is running",
+	s.isRunning = hr.Running
+
+	var summary []interface{}
+	if len(hr.Metrics) == 0 {
+		summary = append(summary, "msg", "sidecar is initializing")
+	} else {
+		summary = append(summary, "msg", "sidecar is running")
+
+		summary = append(summary, hr.MetricLogSummary(config.ProcessedMetric)...)
+		summary = append(summary, hr.MetricLogSummary(config.OutcomeMetric)...)
+		summary = append(summary, hr.MetricLogSummary(config.DroppedSeriesMetric)...)
 	}
-	summary = append(summary, hr.MetricLogSummary(config.ProcessedMetric)...)
-	summary = append(summary, hr.MetricLogSummary(config.OutcomeMetric)...)
-	summary = append(summary, hr.MetricLogSummary(config.DroppedSeriesMetric)...)
 
 	level.Info(s.logger).Log(summary...)
 
@@ -415,11 +400,12 @@ func (s *Supervisor) noteHealthy(hr health.Response) string {
 
 func (s *Supervisor) noteUnhealthy(hr health.Response) string {
 	if hr.Stackdump != "" {
+		s.veryUnhealthy = true
+
 		summary := []interface{}{
-			"msg", "process is very unhealthy",
+			"msg", "process is unhealthy",
 			"stackdump", hr.Stackdump,
 		}
-		s.veryUnhealthy = true
 		level.Warn(s.logger).Log(summary...)
 	}
 	return "unhealthy"